Address code review changes - SQUASH ME?
